requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
telethon==1.36.0
qrcode[pil]==7.4.2
